feat: google drive video stream integration and auto duration calculation

+++ app/Models/Lesson.php
@@ -26,7 +26,7 @@ class Lesson extends Model
     {
         return [
             'order' => 'integer',
-            'duration_minutes' => 'integer',
+            'duration_minutes' => 'float',
         ];
     }
